# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "這是我們的 API 文件",
        "contact": {
            "email": "xxx"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/createinvoiceinfo": {
            "post": {
                "tags": [
                    "base_invoiceinfo"
                ],
                "summary": "新增發票資料",
                "description": "新增發票資料",
                "operationId": "createinvoiceinfo",
                "parameters": [
                    {
                        "name": "period_start",
                        "in": "query",
                        "description": "期別_起",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period_end",
                        "in": "query",
                        "description": "期別_迄",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "invoice_type",
                        "in": "query",
                        "description": "發票類型",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "track_code",
                        "in": "query",
                        "description": "字軌代碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_number",
                        "in": "query",
                        "description": "發票起始號碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "end_number",
                        "in": "query",
                        "description": "發票截止號碼",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "effective_startdate",
                        "in": "query",
                        "description": "適用起始日期",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "effective_enddate",
                        "in": "query",
                        "description": "適用截止日期",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_valid",
                        "in": "query",
                        "description": "是否有效",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "period_start": {
                                            "type": "string",
                                            "example": "2025-01"
                                        },
                                        "period_end": {
                                            "type": "string",
                                            "example": "2025-02"
                                        },
                                        "invoice_type": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "track_code": {
                                            "type": "string",
                                            "example": "AQ"
                                        },
                                        "start_number": {
                                            "type": "string",
                                            "example": "0000000001"
                                        },
                                        "end_number": {
                                            "type": "string",
                                            "example": "0000000050"
                                        },
                                        "effective_startdate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025/01/01"
                                        },
                                        "effective_enddate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025/02/28"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "單據發票失敗"
                    }
                }
            }
        },
        "/api/invoiceInfo2/{period}": {
            "get": {
                "tags": [
                    "base_invoiceinfo"
                ],
                "summary": "查詢特定發票資訊",
                "description": "查詢特定發票資訊",
                "operationId": "getinvoiceinfo",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "description": "期別",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "period_start": {
                                            "type": "string",
                                            "example": "2025-01"
                                        },
                                        "period_end": {
                                            "type": "string",
                                            "example": "2025-02"
                                        },
                                        "invoice_type": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "track_code": {
                                            "type": "string",
                                            "example": "AQ"
                                        },
                                        "start_number": {
                                            "type": "string",
                                            "example": "0000000001"
                                        },
                                        "end_number": {
                                            "type": "string",
                                            "example": "0000000050"
                                        },
                                        "effective_startdate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025/01/01"
                                        },
                                        "effective_enddate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025/02/28"
                                        },
                                        "is_valid": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到發票資訊"
                    }
                }
            }
        },
...
        "/api/invoiceinfo/showconst": {
            "get": {
                "tags": [
                    "base_invoiceinfo"
                ],
                "summary": "列出所有發票字軌需要的常用(下拉、彈窗)",
                "description": "列出所有發票字軌需要的常用(下拉、彈窗)",
                "operationId": "show_invoiceinfo_aLL_const",
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "發票字軌需要的常用未找到"
                    }
                }
            }
        },
...
    },
...
}